backend/requirements.txt
streamlit
pandas
scikit-learn
nltk
numpy
google-generativeai
flask
flask-cors
gunicorn